refactor
change GameFlowController name to SinglePlayerUno

Move SinglePlayerUno back to uno.game
