+++ Makefile
@@ -13,7 +13,7 @@ SOURCES := \
     $(SRC_DIR)/uno/singleplayergame/PlayerManager.java \
     $(SRC_DIR)/uno/singleplayergame/HumanPlayerController.java \
     $(SRC_DIR)/uno/singleplayergame/OutputRenderer.java \
-    $(SRC_DIR)/uno/singleplayergame/GameFlowController.java \
+    $(SRC_DIR)/uno/game/SinglePlayerUno.java \
 	$(SRC_DIR)/uno/game/MultiplayerGame.java \
     $(SRC_DIR)/uno/multiplayergame/PlayerManager.java \
     $(SRC_DIR)/uno/multiplayergame/OutputRenderer.java \
